fix: move `abi` folder at top level
