+++ package.json
@@ -28,7 +28,9 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "files": [
-    "build/"
+    "build/",
+    "abi/",
+    "src/"
   ],
   "dependencies": {
     "ethers": "~5.7.0"
@@ -48,7 +50,7 @@
     "typechain": "^3.0.0",
     "typedoc": "^0.25.8",
     "typedoc-plugin-markdown": "4.0.0-next.53",
-    "typescript": "^4.3.5",
+    "typescript": "^5.1.6",
     "gts": "^5.2.0"
   },
   "peerDependencies": {
